+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   If this option is set to `true`, `tt` will not create sub-directories for runtime artifacts such
   as control socket, pid file and log file. This option affects only single instance applications.
 - An ability to set different directories for WAL, vinyl and snapshots artifacts.
+- ``tt instances`` command to print a list of enabled applications.
 
 ### Changed